refactor config update for CT

+++ docker.local/bin/conductor/rollback_0chain_config.sh
@@ -4,4 +4,4 @@ set -e
 
 echo "rollback 0chain config";
 
-mv docker.local/config/0chain.yaml.bak docker.local/config/0chain.yaml
+[ -e docker.local/config/0chain.yaml.bak ] && mv docker.local/config/0chain.yaml.bak docker.local/config/0chain.yaml || true
